requirements.txt
# --- Core app / config ---
python-dotenv==1.1.0
PyYAML==6.0.2
click==8.3.1
invoke==2.2.1

# --- Database / networking ---
psycopg2==2.9.10
bcrypt==5.0.0
cryptography==44.0.3
PyNaCl==1.6.1
cffi==2.0.0
pycparser==2.23
websocket-client==1.8.0

# --- HTTP / URL tooling ---
requests==2.32.3
urllib3==2.6.1
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
tldextract==5.3.0
requests-file==3.0.1
update-checker==0.18.0

# --- Reddit ---
praw==7.8.1
prawcore==2.4.0

# --- Data / utilities ---
numpy==2.2.6
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
tabulate==0.9.0
tqdm==4.67.1
packaging==25.0
filelock==3.20.0
colorama==0.4.6

# --- Language detection ---
lingua-language-detector==2.1.1

# --- NLP (spaCy) ---
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
blis==1.3.3
catalogue==2.0.10
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.13
murmurhash==1.0.15
preshed==3.0.12
regex==2025.11.3
smart_open==7.5.0
srsly==2.5.2
thinc==8.3.10
typer-slim==0.20.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1

# --- Pydantic ---
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
typing-inspection==0.4.2
typing_extensions==4.15.0

# --- Presidio ---
presidio_analyzer==2.2.360
presidio_anonymizer==2.2.360
phonenumbers==9.0.20

# --- Templating (if used) ---
Jinja2==3.1.6
MarkupSafe==3.0.3

# --- Testing ---
pytest==8.4.2
iniconfig==2.3.0
pluggy==1.6.0
exceptiongroup==1.3.1
tomli==2.3.0

# --- Build tooling ---
setuptools==80.9.0
Pygments==2.19.2
six==1.17.0
